refactor: Move InputType and UiAction to commonsKmm input package
