+++ aggregates/src/commonMain/kotlin/org/dhis2/mobile/aggregates/model/DataElementInfo.kt
@@ -1,5 +1,7 @@
 package org.dhis2.mobile.aggregates.model
 
+import org.dhis2.mobile.commons.input.InputType
+
 internal data class DataElementInfo(
     val label: String,
     val inputType: InputType,
