+++ commonskmm/src/commonMain/kotlin/org/dhis2/mobile/commons/input/InputType.kt
@@ -1,6 +1,6 @@
-package org.dhis2.mobile.aggregates.model
+package org.dhis2.mobile.commons.input
 
-internal sealed class InputType {
+sealed class InputType {
     data object OptionSet : InputType()
     data object Text : InputType()
     data object LongText : InputType()
